Prohibit tbq for get_state and set_state for now
